# src/docs/swagger.yaml
swagger: "2.0"
info:
  description: "Telecom Hackathon | Build Your App 48hrs Game!\n Build a MiniFlow App Backend"
  version: "0.0.1"
  title: "HackathonT3 | UltraNanoFlow"
  contact:
    email: "kimbuz@teco.com.ar"
host: "bkultrananoflow.herokuapp.com"
basePath: "/api/v1"
paths:
...
  /content:
    get:
      tags:
      - "content"
      summary: "Get list of Content Json"
      description: ""
      operationId: "content"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Api for Content List All"
        required: false
      responses:
        "200":
          description: "Json Output"
...